# packages/shared-types/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "闲遇 API",
    "description": "店长、员工、会员和平台后台统一 API",
    "version": "mvp"
  },
  "paths": {
...
    "/admin/configs/{configKey}": {
      "put": {
        "tags": [
          "admin-platform-config-controller"
        ],
        "operationId": "updateConfig",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AdminUserView"
            }
          },
          {
            "name": "configKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/tenants/{tenantId}/freeze": {
      "post": {
        "tags": [
          "admin-tenant-controller"
        ],
        "operationId": "freezeTenant",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AdminUserView"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenants/{tenantId}/extend-trial": {
      "post": {
        "tags": [
          "admin-tenant-controller"
        ],
        "operationId": "extendTrial",
        "parameters": [
          {
            "name": "actor",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AdminUserView"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/tenants": {
      "get": {
        "tags": [
          "admin-tenant-controller"
        ],
        "operationId": "listTenants",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiry",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTenantListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenants/{tenantId}": {
      "get": {
        "tags": [
          "admin-tenant-controller"
        ],
        "operationId": "getTenant",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTenantDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/configs": {
      "get": {
        "tags": [
          "admin-platform-config-controller"
        ],
        "operationId": "listConfigs",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "editableOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListPlatformConfigResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "AdminUserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "UpdatePlatformConfigRequest": {
        "required": [
          "newValue",
          "reason"
        ],
        "type": "object",
        "properties": {
          "newValue": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "FreezeTenantRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean"
          }
        }
      },
      "ExtendTenantRequest": {
        "required": [
          "newTrialEndAt",
          "reason"
        ],
        "type": "object",
        "properties": {
          "newTrialEndAt": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string"
          },
          "internalNote": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseTenantListResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/TenantListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
      "PageResponseTenantListItem": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantListItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TenantListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "trialStartAt": {
            "type": "string",
            "format": "date"
          },
          "trialEndAt": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "statusText": {
            "type": "string"
          },
          "supportWechatId": {
            "type": "string"
          },
          "issuedCards": {
            "type": "integer",
            "format": "int32"
          },
          "deductions": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          },
          "monthlySalesYuan": {
            "type": "number"
          }
        }
      },
      "TenantListResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/TenantStatsResponse"
          },
          "page": {
            "$ref": "#/components/schemas/PageResponseTenantListItem"
          }
        }
      },
      "TenantStatsResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "trialing": {
            "type": "integer",
            "format": "int64"
          },
          "expiring": {
            "type": "integer",
            "format": "int64"
          },
          "frozen": {
            "type": "integer",
            "format": "int64"
          },
          "extended": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseTenantDetailResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/TenantDetailResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
      "BusinessSnapshotView": {
        "type": "object",
        "properties": {
          "issuedCards": {
            "type": "integer",
            "format": "int32"
          },
          "checkins": {
            "type": "integer",
            "format": "int32"
          },
          "deductions": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          },
          "lowBalanceCount": {
            "type": "integer",
            "format": "int32"
          },
          "expiringCount": {
            "type": "integer",
            "format": "int32"
          },
          "monthlySalesYuan": {
            "type": "number"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StoreView": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "businessCategories": {
            "type": "string"
          },
          "serviceTags": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "businessHours": {
            "type": "string"
          }
        }
      },
      "TenantDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "statusText": {
            "type": "string"
          },
          "trialStartAt": {
            "type": "string",
            "format": "date"
          },
          "trialEndAt": {
            "type": "string",
            "format": "date"
          },
          "frozenReason": {
            "type": "string"
          },
          "supportWechatId": {
            "type": "string"
          },
          "store": {
            "$ref": "#/components/schemas/StoreView"
          },
          "snapshot": {
            "$ref": "#/components/schemas/BusinessSnapshotView"
          }
        }
      },
...
      "ApiResponseListPlatformConfigResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformConfigResponse"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
      "PlatformConfigResponse": {
        "type": "object",
        "properties": {
          "configKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "valueText": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "minValue": {
            "type": "integer",
            "format": "int32"
          },
          "maxValue": {
            "type": "integer",
            "format": "int32"
          },
          "impactScope": {
            "type": "string"
          },
          "editable": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  }
}